requirements.txt
feedparser
requests
python-dotenv
BeautifulSoup4
Flask
pymongo
apscheduler